requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.2
daphne==4.1.2
Django==5.1.7
django-cors-headers==4.7.0
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
pytest==8.3.5
pytest-django==4.10.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
service-identity==24.2.0
setuptools==77.0.3
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
zope.interface==7.2
